+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Translations: [简体中文](CHANGELOG.zh.md)
      prompt for upgrade.
 > 3. Reference [《Migration Documentation》](docs/migrate.md) migrating from 3.x to 4.x
 
+# new
+
+* change: Deprecate AsyncImage with keepContentNoneStartOnDraw parameter
+
 # 4.1.0
 
 > [!CAUTION]